website/chatbot/chatbot copie/requirements.txt
alembic==1.4.1
asgiref==3.5.0
azure-ai-formrecognizer==3.2.0b2
azure-common==1.1.28
azure-core==1.22.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cloudpickle==2.0.0
cryptography==36.0.1
databricks-cli==0.16.4
dill==0.3.4
Django==4.0.2
docker==5.0.3
docopt==0.6.2
elastic-apm==6.7.2
elasticsearch==7.10.0
entrypoints==0.4
farm-haystack @ git+https://github.com/deepset-ai/haystack.git@abc1057869d72b7ce654f40f3945dd2a387aa127
filelock==3.5.1
Flask==2.0.3
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
inflect==5.4.0
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
langdetect==1.0.9
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mlflow==1.13.1
mmh3==3.0.0
more-itertools==8.12.0
msrest==0.6.21
networkx==2.6.3
nltk==3.7
num2words==0.5.10
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prometheus-client==0.13.1
prometheus-flask-exporter==0.18.7
protobuf==3.19.4
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-docx==0.8.11
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
quantulum3==0.7.10
querystring-parser==1.2.4
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tabulate==0.8.9
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.10.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.13.0
typing_extensions==4.1.1
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.3
